pyproject.toml
[project]
name = "yt-interactive-downloader"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "yuta0306", email = "yuta.20010306@gmail.com" }
]
dependencies = [
    "yt-dlp>=2023.7.6",
    "flask>=2.3.2",
    "ffmpeg-python>=0.2.0",
    "tqdm>=4.66.1",
    "requests>=2.31.0",
    "types-requests>=2.31.0.2",
    "python-dotenv>=1.0.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=23.7.0",
    "mypy>=1.5.1",
    "isort>=5.12.0",
    "flake8>=5.0.4",
    "pytest>=7.4.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.rye.scripts]
app = { call = "src.yt_interactive_downloader.frontend.app:run" } 
